chore(dashboard): move validation error message to the body component
