+++ apps/dashboard/src/components/workflow-editor/steps/in-app/in-app-editor.tsx
@@ -4,8 +4,6 @@ import { Notification5Fill } from '@/components/icons';
 import { Separator } from '@/components/primitives/separator';
 import { getComponentByType } from '@/components/workflow-editor/steps/component-utils';
 import { InAppTabsSection } from '@/components/workflow-editor/steps/in-app/in-app-tabs-section';
-import { FormMessagePure } from '../../../primitives/form/form';
-import { useFormContext } from 'react-hook-form';
 
 const avatarKey = 'avatar';
 const subjectKey = 'subject';
@@ -15,9 +13,6 @@ const primaryActionKey = 'primaryAction';
 const secondaryActionKey = 'secondaryAction';
 
 export const InAppEditor = ({ uiSchema }: { uiSchema?: UiSchema }) => {
-  const { getFieldState } = useFormContext();
-  const bodyError = getFieldState(bodyKey).error;
-
   if (!uiSchema || uiSchema?.group !== UiSchemaGroupEnum.IN_APP) {
     return null;
   }
@@ -45,14 +40,7 @@ export const InAppEditor = ({ uiSchema }: { uiSchema?: UiSchema }) => {
               {subject && getComponentByType({ component: subject.component })}
             </div>
           )}
-          {body && (
-            <>
-              {getComponentByType({ component: body.component })}
-              <FormMessagePure
-                error={bodyError ? String(bodyError.message) : undefined}
-              >{`Type {{ for variables, or wrap text in ** for bold.`}</FormMessagePure>
-            </>
-          )}
+          {body && getComponentByType({ component: body.component })}
           {(primaryAction || secondaryAction) &&
             getComponentByType({
               component: primaryAction.component || secondaryAction.component,
